Revert to 'expovariate' for voter arrival; use Excel file's arrival rate
As it turns out, using a Poisson distribution is incorrect for our application. It returns the number of occurances of a given event but not the time of the event (which is what we need). So this commit reverts to the 'expovariate' function which generates arrival times that fit a Poisson distribution. This commit also uses the Excel file's arrival rates.