Merge voter_sim.py changes into main (#22)
* Add note about 'arrival_rt' recalcuation within 'voter_sim.py'

* Calculate voter arrival times with 'random.poisson' instead of 'expovariate'

This change has some negative side effects including: (1) increasing randomness; (2) slightly increasing run-time; (3) breaking the calculation for Location #2.

* Revert to 'expovariate' for voter arrival; use Excel file's arrival rate

As it turns out, using a Poisson distribution is incorrect for our application. It returns the number of occurances of a given event but not the time of the event (which is what we need). So this commit reverts to the 'expovariate' function which generates arrival times that fit a Poisson distribution. This commit also uses the Excel file's arrival rates.

* Remove arrival rate calculation from 'fetch_location_data.py'

To simplify the code, this commit moves the calculation for 'arrival_rt' into 'voter_sim.py'. It also removes the passed-in value from the 'voter_sim' function and 'VotingLocation' constructor.

* Remove unused 'Settings' import from 'izgbs.py'

* Add a 'requirements.txt' file specifically for Github Actions unit tests

* Update Github actions file to use new 'requirements.txt' file

This requirements file only includes the packages necessary to run the unit tests.